Added battery bot and saw
I have added the prefabs for both the saw and the battery bot. Since
they currently use the same script, I have included that as well. The
sprite probably won't render since I used some art since I made myself.
The script is very similar to the simple moving platform script with the
addition of a function that resets the player when they trigger the
collision. The trigger function is very easy to change if we would
rather use checkpoints or have damage. Currently, all they do is just
move back and forth and if you touch them it resets the player. I
figured since we wanted them to be like a Goomba, this would fit that. I
still would like to add some extra functionality and interaction with
the magnet gun though. Hopefully I will be able to do that before this
iteration's end.