Clamping
I've fixed clamping by increasing the clamp detection distance from 0.65 to 1 (line 157).
On lines 262 - 263 and 267, I made it so Maggie will freeze while clamped. Her arm can still move though, and this needs to be fixed.
On line 342 - 345, I added the ability for Maggie to jump while clamped.

Also, I found a potential bug on line 182, and added a comment to warn of this.