Battery bot and saw changes
Justin Tomlinson
CEN4020-02
Iteration 2

The battery bot and saw are now made up of two separate scripts.
The movement of the battery bot and saw have been completely
reworked. Previously the movement was very finicky and hard to work
with. Now there are no positions that need to be entered. You simply
place the object down and choose the distance that it goes (which
applies for both directions because where it is placed is its center.
Additionally, there is now a static battery bot which does not move and
can ride on Cody's moving platforms. Finally, the saw will move in the
same way as the battery bot unless it feels the attraction of Maggie's
magnet, in which case it will move towards the most recent attraction
that it felt.
Also added art for the battery bot and saw.